Updated Arun's Merge by formatiing as required. Camelcase and Floating point as required
